Add files via upload
This projects aims to build a Machine Leaning model that can be used to predict the Poverty Level (GINI Index) of a country based on the international debt.  Data consist of the sets of data from the world bank data bank. The Intl debt data consist of about 400,000 rows of data and about 150 countries.

Data was cleaned and formatted to fit for model development.